refactor(core): 将 core 中的测试文件和文档文件归至 tests 和 docs 文件夹中
