+++ src/core/components/docs/active-object-document.md
