+++ src/core/components/docs/board-manager-document.md
