+++ src/core/components/docs/page-load-manager-document.md
