+++ src/core/components/docs/page-manager-document.md
