+++ src/core/components/docs/page-object-manager-document.md
