+++ src/core/components/docs/tier-graph-document.md
