+++ src/core/components/tests/aom/page-load-manager.mock.js
